+++ piecrust-uplink/src/abi/handlers.rs
@@ -4,14 +4,27 @@
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
+use core::fmt::Write;
 use core::panic::PanicInfo;
 
+extern "C" {
+    pub fn panic(arg_len: u32);
+}
+
 #[panic_handler]
-fn panic(_info: &PanicInfo) -> ! {
-    #[cfg(feature = "debug")]
-    if let Some(msg) = _info.message() {
-        crate::debug!("{msg}");
+unsafe fn handle_panic(info: &PanicInfo) -> ! {
+    let mut w = crate::ArgbufWriter::default();
+
+    // If we fail in writing to the argument buffer, we just call `panic` after
+    // writing a standard message instead.
+    if let Some(args) = info.message() {
+        if w.write_fmt(*args).is_err() {
+            w = crate::ArgbufWriter::default();
+            let _ = write!(w, "PANIC INFO TOO LONG");
+        }
     }
+
+    panic(w.ofs() as u32);
     unreachable!()
 }